Move `Terminal` trait to its own file
This needs to be done because the new encrypted custody backend will
also make use of the trait, so it needs to be moved outside of that
particular module.
