+++ crates/custody/src/lib.rs
@@ -13,6 +13,7 @@ extern crate serde_with;
 mod client;
 mod pre_auth;
 mod request;
+mod terminal;
 
 pub mod null_kms;
 pub mod policy;
